kvm: pkvm: Refactor pkvm_shadow_vcpu_handle and shadow_vm_handle
The pkvm_shadow_vcpu_handle was defined in kvm_vcpu to represent a vcpu
in the pkvm hypervisor which is available to x86/arm only. Move this to
kvm_vcpu_arch. This is similar to shadow_vm_handle which is defined in
kvm_arch. Meanwhile align their name as pkvm_vcpu_handle and
pkvm_vm_handle.

Previously the pkvm_shadow_vcpu_handle was constructed by the
shadow_vm_handle in the high 32bits and the vcpu index in the low
32bits. To simplify, redefine the pkvm_vcpu_handle as an int which only
contains the vcpu index.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>