pkvm: vmx: Enhance nested_flush_shadow_ept to support guest vcpu
The current nested_flush_shadow_ept assumes the input parameter struct
kvm_vcpu pointer always representing the vcpu of the host. Enhance this
function by getting the pkvm_host_vcpu via the CPU number in the
kvm_vcpu. With this, the input parameter can be a guest kvm_vcpu
pointer.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>