pkvm: io: Refactor the return value for try_emul_host_mmio
Current try_emul_host_mmio returns different errno, indicating quite
different processing path in caller side. Particularly, -EINVAL means
the MMIO region is not in emulated handler list, and the caller will
setup EPT for the MMIO and let the host_os retry. While other return
value (even it is an error number) means it is handled (with error).
This is hard to read.

Change try_emul_host_mmio to return true/false indicating if the MMIO
is emulated or not. If it is emulated (even if emulating using
default handler or with error number), the caller is all set. If it
is not emulated, the caller can then setup EPT for the MMIO and let
the host_os retry.

Signed-off-by: Eddie Dong <eddie.dong@intel.com>