pkvm: vmx: Handle force_immediate_exit
Make the preemption timer as the must have feature for the pkvm
hyperviso so that this can be used to support the force_immediate_exit.
With preemption timer enabled, the pkvm hypervisor can use preemption
timer to implement a force immeidate exit instead of relying on the host
to send an ipi. For the vmexit due to the force_immediate_exit requested
by the host KVM, breaks the pkvm vcpu run loop and switches back to the
host.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>