kvm: pkvm: Implement check_emulate_instruction
The check_emulate_instruction callback can be used by the host KVM
during emulating a MMIO instruction. For the npVM, returning
X86EMUL_CONTINUE is good enough. For the pVM, this is not expected if
the pVM is enlightened by using hypercall to access MMIO. And if the pVM
still expects receiving #VE, returning X86EMUL_RETRY_INSTR can let the
pVM to retry after clearing the suppress #VE bit in the EPT.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>