pkvm: x86: Add kvm_x86_vendor_init
Add kvm_x86_vendor_init function for the pkvm hypervisor which does the
similar thing as for the linux KVM. As most of the linux KVM code in
this function cannot be used by the pkvm hypervisor, this function is
mostly rewritten by doing the necessary initialization for the pkvm
hypervisor.

As some of the code/data is annotated as the __init/__initdata, add
init.text and init.data sections to the pkvm hypervisor.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>