pkvm: x86: Introduce pkvm_x86_ops
Introduce pkvm_x86_ops which contains pkvm specific operations. The VMX
can implement its pkvm_x86_ops for the pkvm x86 core code to use. The
switch_to_guest_vcpu/host_vcpu callbacks are defined in this ops which
are used by the pkvm hypervisor to switch the vcpu context between the
guest and the host.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>