pkvm: x86: Add kvm_vcpu_reset
Add kvm_vcpu_reset and the dependency functions by porting from the linux
KVM x86 code. Some vcpu reset code is disabled in the pkvm hypervisor
due to the corresponding features are not supported, e.g., the guest fpu
state is still managed by the host and the PV features are not
supported.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>