pkvm: vmx: Implement vcpu_pre_run
With the commit be6802a642c5 "pkvm: x86: Enforce pvmfw as the pVM entry
point", the RIP register on the first running vCPU is enforced as the
pvmfw address in the VMX emulate-based method. The similar thing should
be done in the PV-based method, so enforce the pvmfw as the pVM entry
point via the vcpu_pre_run.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>