pkvm: vmx: Disable EPT AD bits feature
The EPT is shadowed by the pkvm hypervisor. So the EPT managed by the
host is a virtual EPT which doesn't have AD bits feature. Set the
enable_ept_ad_bits as false from both the pkvm hypervisor and the host
side to make this consistent.

This should be revisited when the PV EPT method is ready.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>