kvm: pkvm: Use PV interface to access guest MSRs
Add pkvm_emulated_msr to check if a MSR is emulated by the pkvm
hypervisor or not. The list will be extended if more MSR can be
emulated by the pkvm hypervisor. If an MSR is emulated by the pkvm
hypervisor, use the PV interface to access. Otherwise the host KVM
should emulate.

Once the guest state is protected, the pkvm hypervisor emulated guest
MSRs are not allowed to be accessed via these PV interfaces.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>