pkvm: x86: Handle EFER MSR vmexit
Add set_efer to support emulating reading and writing EFER MSR. With
modifying the KVM_MMU_EFER_ROLE_BITS, the mmu is required to be reset.
Currently the mmu reset is done by the host via the returning request
HOST_RESET_MMU in the pkvm_run page. This can be revisited later when PV
EPT is ready.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>