kvm: pkvm: Use PV interfaces to load and put a vcpu
Use PV interfaces to load and put a vcpu. When the host KVM loads a vcpu
on a perticular CPU, the vcpu_put PV interface will be used to do a vmcs
clear if this vcpu is loaded on another CPU before and then use the
vcpu_load PV interface to load the cleared VMCS. When the host KVM puts
a vcpu, the vcpu_put PV interface is not used to do the VMCS clear as
the same vcpu might be loaded again on the same CPU.

Most of the vcpu_load/put logic are done in the pkvm hypervisor but the
Posted-Interrupt load and put are left to the host KVM, which should be
revisited later to see if and how this could be done in the pkvm
hypervisor.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>