pkvm: vmx: Use vmcs_store to get the active VMCS physical address
With adding more VMX code in the pkvm hypervisor, vmcs_load is actively
used to load VMCS page. However the pkvm hypervisor wishes to use the
vmcs_load_track function which tracks the current active vmcs via a
software method, so that it can know which VMCS should be re-actived
after switch to the host vcpu VMCS to enable the NMI/IRQ window by the
NMI handler.

With the vmcs_store inline function, the current active VMCS physical
address can be got via the vmptrst instruction so no need to rely on the
software. Change to use vmcs_store so that the added VMX code can
continue to use the vmcs_load (instead of vmcs_load_track).

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>