REVERTME: pkvm: vmx: Sync vmcs to satisfy the mixed VMX usage
As the VMX emulation method is still in use and the guest vmcs is used
as shadow vmcs of the host vcpu vmcs, when switches from the host vcpu
to the guest vcpu, the guest vmcs is loaded with its shadow_vmcs
indicator disabled, and the cached emulated fields in the vmcs12 should
be synced to the guest vmcs.

Similarly, when switches from the guest vcpu to the host vcpu, the
emulated fields should be synced from the guest vmcs to the vmcs12 to
satisfy the VMX emulation method, and then enable the shadow_vmcs
indicator in the guest vmcs to use it as the shadow vmcs of the host
vmcs.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>