REVERTME: pkvm: vmx: Overwrite the host RIP
As the VMX emulate-based method is still in use, the HOST_RIP is set by
the pkvm_nested_load_vmcs() as __pkvm_vmexit_entry. With the PV-based
method, the HOST_RIP will be the vmx_vmexit which should be configured
by the vcpu_reset. Before the vcpu_reset is implemented in the pkvm
hypervisor, hack by overwriting the HOST_RIP as the vmx_vmexit in the
vmx_vcpu_pre_run.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>