pkvm: x86: Add vcpu_run PV interface
Add vcpu_run PV interface for the host KVM to run a guest vcpu. The
guest vcpu runs in a loop which can re-enter the guest vcpu without
going back to the host KVM if the pkvm hypervisor allows. The context
switching between host vcpu and the guest vcpu will be done in the
following patch.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>