REVERTME: pkvm: vmx: Sync vcpu_vmx structure between private and shared
currently the VMX is managed by both the host KVM and the pkvm
hypervisor, which requires the key data structure vcpu_vmx keeping in
sync. Doing the sync during switching vcpus. When switches to the guest
vcpu, copy the necessary data from the shared vcpu_vmx to the private
vcpu_vmx. And when switches to the host vcpu, copy the necessary data
from the private vcpu_vmx to the shared vcpu_vmx.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>