pkvm: x86: Handle the KVM_REQ_EVENT request
With the capability of injecting irq/nmi/exception in the pkvm
hypervisor, the KVM_REQ_EVENT can be handled by injecting the pending
events. Before entering the guest, do the event injection if there is
pending KVM_REQ_EVENT.

During injecting events, the pkvm hypervisor will check if there is any
pending exception but won't check for the irq/nmi as they are managed by
the host KVM. For any not-allowed pending events, req_immediate_exit
flag will be set to trigger the guest vmexit immediately, so that the
pkvm hypervisor can re-inject the pending events in the next cycle.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>