pkvm: vmx: Share pVM vcpu state to the host for emulating MSR
Share the guest RAX/RDX/RCX registers to the host for the rdmsr/wrmsr
emulation.

Add the complete_emulated_msr PV interface for the host to send the msr
emulation error code to the pkvm hypervisor. For the pVM, the pkvm
hypervisor will check the last vmexit reason and complete the MSR
emulation(either injecting #GP or skipping the instruction) according to
the error code before entering the guest. For the npVM, the pkvm
hypervisor can directly complete the MSR emulation.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>