kvm: pkvm: Add PV interface to get/set nmi mask
Add two new PV interfaces for the host VMM to get and set nmi mask. The
implementation is similar with the linux VMX code. Once the guest state
is protected, the nmi mask is not allowed to be set via the set PV
interface. While the nmi mask is always allowed to be got via the get PV
interface as the host KVM will need this for injecting NMI.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>