kvm: pkvm: Handle the uret MSRs switching
The prepare_switch_to_guest/host have been implemented in the pkvm
hypervisor, which are used before entering into the guest and after
breaking out the vcpu run loop. The host KVM doesn't need to explicitly
use the PV interface to explicitly call prepare_switch_to_guest/host.
Although the uret MSRs are part of the switching, the emulation are
still done by the host KVM. Due to this, the uret MSRs switching are
left to the host KVM. The switching and emulation will be moved to the
pkvm hypervisor together.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>