pkvm: vmx: Deliver pVM's cr0/cr4/efer to the host KVM to init/reset MMU
The requests HOST_INIT/RESET_MMU are handled by the host KVM to init or
reset the mmu which requires to detect the guest mmu attributes. Such
attributes are determined by the vcpu CR0/CR4/EFER registers. For the
npVM, these registers are already shared to the host unconditionally. So
only deliver these for the pVM when sends such requests to the host KVM.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>