pkvm: vmx: Support guest single-setup debug exception
With TF bit is set in guest rflags, executing instruction in the guest
can cause exception_nmi vmexit if the DB bit is set in the exception
bitmap. To make the guest can receive the DB exception, this would be
done by either removing the DB bit from the exception bitmap or
injecting the DB exception from the exception_nmi vmexit handler. To
keep align with the linux VMX implementation, choose the second option.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>