pkvm: x86: Sync debug registers for the npVM
To support debugging guest from the host side, the host may change the
guest debug registers. Meanwhile the pkvm hypervisor handles the
DR_ACCESS vmexit. So the debug registers will be accessed by the both
sides, which requires to keep the guest debug related registers in sync.
For the security reason, this is only allowed for the npVM.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>