pkvm: x86: Enhance the debug registers switching
Before entering the vcpu run loop, save the current debug registers for
the host as the debug registers may be overwriting with the guest debug
registers. And restoring those debug registers after exiting the vcpu
run loop.

The guest debug registers are loaded before vcpu running if it is
necessary. And syncing the dirty debug registers for the guest in the
vcpu run loop so that the up-to-date dr variables in the vcpu->arch can
be referenced.

Fixes: e2f430e2b53c ("pkvm: x86: Switch the debug registers")
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>