pkvm: x86: Fix occasional compiling failures for the kvm-asm-offsets.h
The kvm-asm-offsets.h in kvm folder is required by both the kvm/vmx and
pkvm/vmx. The current implementation is that both the Makefiles are
generating the kvm-asm-offsets.h in the kvm folder, which can cause
occasional compiling failures due to conflicts. To fix this, generating
the kvm-asm-offsets.h in the kvm/pkvm folder for the pkvm hypervisor and
use this header file as its own dependency. Also update the .gitignore
file to ignore the all the kvm-asm-offsets.h files in the kvm folder.

Fixes: 3729703bb1fa ("pkvm: vmx: Implement vcpu_run")
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>