pkvm: x86: add assembler workaround
So that old GNU assembler can also work.

The workaround is to generate assembly code first from .c, then patch the code
to add section name suffix with current line number, so that .pushsection always
uses a unique name.

It seems that CLANG by default will convert .pushsection in inline assembly
code to .section, which makes the workaround not working. Luckily CLANG has
parameter of "no-integrated-as" to disable it.

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>