kvm: Add #ifndef __PKVM_HYP__ for pkvm to use
This change is to let the pkvm hypervisor reuse the KVM's source
code files (x86.c, cpuid.c, lapic.c vmx.c, vmenter.S) rather than having
its own files via manutally adding the check "#ifndef __PKVM_HYP__" in
the KVM's source code file. This can remove the unused code for the pkvm
hypervisor to make the compiling successful.

The topic/use-kvm-files-directly-in-pkvm is for the same purpose which
is via using linker to remove. It will be a preferred way when it is
ready.

The purpose of doing so is to understand the code maintain burden during
upgrading the kernel version.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>