ANDROID: pkvm: x86: Add the fpu register swapping functions
Add the fpu_swap_kvm_fpstate() for the pkvm hypervisor to swap the FPU
registers to guarantee the FPU isolation. The implementation is ported
from the linux kernel source code with minor modifications that, only
saving/restoring the FPU registers for the pVM, and wipe the FPU
registers by restoring with the initial FPU state before returning to
the host. For the npVM, no FPU registers swapping.

Bug: 420676925
Upstream-Task: 402758258

Change-Id: I6463cf9b2b7dcb9251016f8490db35fcde775ee0
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>