pkvm: vmx: Add initial pkvm host support
Introduce the CONFIG_PKVM_INTEL for enabling the pkvm hypervisor support
on Intel platform. With CONFIG_PKVM_INTEL=y, the vmx driver will try to
initialize the pkvm driver if the vmx is supported by the hardware. In
the initialization, the pkvm driver will create a non-root mode
environment via the vmx and deprivilege the linux kernel which runs as
the host VM. The pkvm hypervisor code will run in the root mode to
handle the vmexits from the host VM.

The current implementation is just simply initializing the pkvm_hyp
structure which will be extended by the subsequent patches to implement
the detailed logic.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>