ANDROID: pkvm: x86: Add host memory pages sharing/unsharing with pKVM
Add two new page state APIs to share/unshare host memory pages with the
pKVM hypervisor.

With sharing a host memory page with the pKVM hypervisor, the host page
state is transferred from OWNED to SHARED_OWNED. After unsharing, the page
state is transferred from SHARED_OWNED to OWNED.

These two APIs will be exposed to the host via new PV interfaces so that
the host can lead this sharing/unsharing by the subsequent patch when it
is necessary.

Upstream-Task: 402758258

Fixes: ed3834af0771 ("ANDROID: kvm: pkvm: Use PV interface for vm/vcpu life cycle management")
Change-Id: Ia321be7bdd5b23befbf00290436683a5d5d9ecc7
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>