pkvm: x86: Create percpu for the pkvm hypervisor
The percpu mechanism is useful for the pkvm hypervisor to manage percpu
data. However the pkvm hypervisor cannot use the linux kernel's percpu
memory for security reason. And having pkvm's own percpu memory is also
helpful for the pkvm to reuse some of the linux kernel's code.

Create pkvm's percpu section so that all the pkvm defined percpu will be
located in this section. Reserve more memory pages for each CPU which
will be used as the __per_cpu_offset and this_cpu_off.

The __per_cpu_offset will be used as pkvm's GS_BASE when enabling the
isolation by the later patches.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>