pkvm: x86: Reuse retpoline.S in kernel for pkvm
Copy the retpoline.S in the kernel as a new file retpoline.pkvm. S, and
rename the defined symbols manucally by adding with '__pkvm' suffix. The
retpoline.pkvm.S will be compiled as the pkvm hypervisor's own retpoline
module. To satisfy this, add retpoline.S required symbols to the pkvm
hypervisor and disable the unnecessary ones.

The reason to manucally rename is that the retpoline.S itself has the asm
code "jmp __x86_return_thunk" which will be patched by the kernel at the
boot time as well. So this symbol name should be retained. However the
__x86_return_thunk function is also defined in retpoline.S and this
function should be renamed to avoid the symbol naming conflict.

[chuanxiao: add compiling command]

Signed-off-by: Zhiyuan Lv <zhiyuan.lv@intel.com>
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>