pkvm: vmx: Handle the deprivileged host vmexit
Although the VMCS is configured to minimize the vmexits for the
deprivileged host, there are still some vmexits cannot be avoid. The
CPUID/XSETBV vmexit is handled by re-executing the same instruction on
the native and returning the corresponding result to the host. The CR
vmexit only handles mov-to-cr4 to prevent the host really controlling
the VMX on hardware. The MSR vmexits are handled by ignoring the invalid
MSR accessing. The other types of vmexits should not happen at this
moment.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>