pkvm: vmx: Handle NMI
NMI should be handled by the deprivileged host, while it may happen when
the pkvm hypervisor is running. To make sure such NMI can be injected to
the host, register a NMI handler to capture the NMI. The NMI handler
will set the pending_nmi flag and enable the irq window for NMI
injection in case the NMI is captured late. The captured NMI will be
injected to the host before vmenter.

To handle the EXIT_REASON_INTERRUPT_WINDOW, add a new vmexit handler for
the host.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>