pkvm: x86: Add pkvm_pcpu instance to present phys CPU
Add structure pkvm_pcpu for the pkvm hypervisor to represent the phys
CPU. Each phys CPU will have a corresponding pkvm_pcpu instance in the
pkvm hypervisor, which requires to reserve addtional memory. Stack and
is the elements in the pkvm_pcpu which will be used as the RSP for the
pkvm hypervisor.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>