pKVM: VMX: Handle NMI
NMI should be handled by the deprivileged host, while it may happen when
the pKVM hypervisor is running. To make sure such NMI can be injected to
the host, register a NMI handler to capture the NMI. The NMI handler
will record the pending NMI and request an immediate vmexit from the
host VM which is implemented via using preemption timer, to inject in
case the NMI is captured late. The captured NMI will be injected to the
host before vmenter. If the NMI is blocked, then open the irq window to
inject when the NMI is unblocked.

If an exception is injected, will also request an immediate vmexit from
the host VM, and then inject the NMI before the next vmenter.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>