pKVM: x86: Add pkvm_hyp instance
Add pkvm_hyp instance to represent the pKVM hypervisor. The num_cpus and
pcpus array are the only elements at the moment, which represents the
possible physical CPUs. Each pkvm_pcpu has a 16K u8 array as the stack of
the pKVM hypervisor on each physical CPU. The linux kernel will reserve
enough memory pages for these instances according to the possible CPU
number.

The pkvm_hyp and pkvm_pcpu instances are allocated via the early alloc
APIs from the reserved memory, and complete the basic initialization
during the pKVM init phase.

As the reserved memory will not be released back to the linux, its error
should not impact the initialization of the KVM, so will let the KVM
continue init.

Squash changelog:
  c1a901671fde ("pKVM: x86: Index host_vcpus and pcpus via consecutive number")

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>