KVM: pVMX: Add initial pKVM host interface
Add the initial implementation of the pKVM host, which serves as a
middle layer between the pKVM VMX host and the pKVM hypervisor. An empty
pkvm_host_x86_ops structure is introduced, which will be extended to
provide the necessary operations for the pKVM VMX host to interact with
the pKVM hypervisor to run guest VMs.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>