pKVM: x86: Introduce struct pkvm_memcache
Introduce struct pkvm_memcache, which can be used to store a list of
page-aligned physical memory ranges.

The helper push_pkvm_memcache() is to store a memory range's physical
address and size in the pkvm_memcache.

The helper pop_pkvm_memcache() is to retrieve the last stored memory
range's physical address and size from the pkvm_memcache.

Squash changelog:
  40f26d2b4b3f ("pKVM: x86: Reject zero-size push in push_pkvm_memcache()")
  f7722d51fe3e ("pKVM: x86: Move memcache helpers to kvm_pkvm.h")

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>