pKVM: x86: set PKVM_INTEL depend on !BLK_DEV_FD
To avoid the INTEL_IOMMU_FLOPPY_WA workaround of mapping first 16MB to a
iommu domain, disable PKVM_INTEL when BLK_DEV_FD is enabled.

INTEL_IOMMU_FLOPPY_WA was introduced to guarantee successful
functionality of floppy devices when iommu is enabled. floppy driver
has not yet adapted the latest dma apis and hence does not go through
the dma/iommu layer to reserve the dma pages. As a result, floppy device
dma will fail as the pages are not mapped in iommu domain page tables.
INTEL_IOMMU_FLOPPY_WA maps the first 16MB memory in the iommu domain
page tables as a work around to let the dma operation go through.

This creates issues with pkvm. pkvm guarantees that dma pages(pages
mapped in iommu domain page tables) cannot be donated. This is to avoid
donated pages being accessed by devices. This 16MB mapping created by
iommu driver stays as long as the domain is active, but the floppy driver
may be using only a portion of it or might not be using it at all. The host
kernel can and will use this memory if free (espescially during low
memory situations). Host might donate pages from this 16MB mapped range
and pkvm needs to verify that this page does not belong to a dma region.
But pkvm doesn't have information about whether the floppy driver is using
the memory or not and hence rightfully errors out when pages from this
16MB range is donated.

Floppy driver cannot succesfully operate without INTEL_IOMMU_FLOPPY_WA
when iommu is enabled and pkvm has a hard dependency on iommu, so the
floppy driver and pkvm are incompatible. This patch makes it explicit.

Dmytro: since v6.18 there is one more reason to ensure disabling
INTEL_IOMMU_FLOPPY_WA: pKVM for simplicity requires each DMA-mapped
range to be either fully normal memory or fully reserved memory, not a
mix of both, while those first 16MB are almost always a mix.

Squash changelog:
  e9f1086a17b7 ("SQUASHME: pKVM: x86: Fix dependency on !BLK_DEV_FD")

Signed-off-by: Vineeth Pillai <vineethrp@google.com>
Signed-off-by: Dmytro Maluka <dmaluka@google.com>