ANDROID: pkvm: x86: Warn if root_to_sp() is called
With pKVM's PV MMU, mmu->root.hpa does not represent an actual physical
address of a shadow page (as there are no shadow pages) and thus calling
root_to_sp() is incorrect. Code paths that call it should not be
triggered if pKVM is enabled. Add a WARN_ON_ONCE so that if it is still
triggered for some reason, we'll notice that.

Bug: 391541180
Test: run various VMs, including KVM unit tests and selftests, and check
if no call trace in dmesg.
Upstream-Task: 402758258

Change-Id: I480b879561b6b476e0a9d324fd65fce39416372c
Signed-off-by: Dmytro Maluka <dmaluka@google.com>