ANDROID: pkvm: x86: Don't allow host to flush TLB of a running pVM
Although the host still needs to request pKVM to flush TLB for a VM in
some cases:

- Flush stage-2 (i.e. EPT) TLB for npVM after asking pKVM to clear the
  access bit in an EPT PTE, in the .clear_flush_young() MMU notifier
  (as opposed to .clear_young() which does not require flushing TLB; so
  in this case it is the host that decides whether to flush EPT TLB or
  not).

- Flush VPID-tagged TLB (and in some cases EPT TLB as well) after
  changing CR0/CR3/CR4/EFER values in SET_SREGS ioctl or when emulating
  instructions for npVM.

all these cases either apply to npVM only, or apply to pVM as well but
only if the given vCPU of a pVM has not started running yet.

So make the TLB flushing PV interfaces inaccessible for a protected vCPU
that has already started running, to avoid letting the host affect the
pVM's internal state without a real need.

Bug: 391541180
Test: run various VMs
Upstream-Task: 402758258

Change-Id: I6f81cad822c79331af6f35fa64f8d88825efd128
Signed-off-by: Dmytro Maluka <dmaluka@google.com>