ANDROID: pkvm: x86: Allocate the fpstate for the guest vcpu
The pkvm hypervisor needs to save/restore the FPU registers to/from the
memory for the guest to support the FPU isolation. The fpstate memory
is allocated by the host and send to the pkvm hypervisor via the
vcpu_create PV interfaces, which will be donated to pkvm hypervisor for
protection. It will be undonated when destroy the VM.

For the pVM, as the fpstate memory will be used by the pkvm hypervisor
to save/restore the pVM's FPU registers for the isolation purpose, its
size is counted according to the guest fpu allocated by the host.

For the npVM, this fpstate memory doesn't have to be used in the same as
no FPU isolation required. Instead, the fpstate memory owned by the host
can be used to save/restore the FPU registers for the npVM. So the size
for this fpstate memory is just size of the structure fpstate except for
the regs, which is large enough for the pkvm hypervisor to emulate the
IA32_XFD msr.

Bug: 420676925
Upstream-Task: 402758258

Change-Id: I5aa9920f78b49fc2196d8ff5e506808d7da44fd9
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>